Шаблонизируй это: 1. принимаем данные
